multi_image: creation of domain image
Adding the possibility of creating domain images.

A domain image is similar to a child image in terms of CMake configure,
generator and build steps.

The main difference is that the hex file for a domain image is not
merged into the parent hex image.

Adding a domain image will create a dedicated flash target.
This domain image flash target will be executed as part of the parent
image flash target.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>