cmake: now using external project for edge impulse compiling
Using ExternalProject for building the edge impulse library.

This allows downloading and extracting the zip file containing the model
on each build invocation.

In the event that sources are added or removed from the library the
ExternalProject ensures that the build system will correctly handle this
in the build system.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>