bluetooth: rpc: fixup on Bluetooth RPC Konfig file
BT_RPC_CLIENT and BT_RPC_HOST both selects BT symbol.

This is fundamentally wrong, as BT_RPC is a feature that is depending
on Bluetooth being enabled in the system.

In such cases, a `depends on BT` is more correct than selecting BT.

Changing from a select to a depends also fixes circular issues and
allows BT_RPC to only be enabled for devices that are not using
features not currently supported by the RPC, for example GATT.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>