samples: radio_test: relocate dts-nrf21540-fem.overlay to app root
The configuration folder is only used for applications where there is
a wish to structure the more complex configurations for each supported
board in a dedicated folder. A folder that is named after the board
itself.

As this is not the case for the radio_test sample, then the
dts-nrf21540-fem.overlay has been moved to the sample root folder so
that it is aligned with how samples are normally structured.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>