mcuboot: cmake: BOOT_IMAGE_ACCESS_HOOKS_FILE removed in favor of CMake
The BOOT_IMAGE_ACCESS_HOOKS_FILE has been removed in favor of CMake.

The setting BOOT_IMAGE_ACCESS_HOOKS in Kconfig is still used to inform
if support for an external hook file should be enabled, but the hook
file itself is now added to the mcuboot_util library in the
modules/mcuboot/CMakeLists.txt file.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>