cmake: Support specifying Zephyr version with find_package(Ncs)
The Zephyr CMake package is automatically sourced when using
find_package(Ncs) as Zephyr is a required component.

This commit provides the user the ability to request a specific Zephyr
version by setting the COMPONENTS argument like this:
> find_package(Ncs 1.7.0 REQUIRED COMPONENTS Zephyr 1.6.99)

which will request Ncs to use Zephyr version 1.6.99.
If an exact match is required, user can do:
> find_package(Ncs 1.7.0 EXACT REQUIRED COMPONENTS Zephyr 1.6.99)

which will be propagated to the Zephyr packgae like:
> find_package(Zephyr 1.6.99 EXACT REQUIRED)

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>