[wip] signing: specify pad-header and fixed header size
Specify `--pad-header` for west sign, this is required when the image
header is not pre-filled with zeros.
Vanilla Zephyr specifies ROM_START_OFFSET which ensures the header is
pre-filled with zeroes, but with partition manager, the image starts
at the defined rom address, hense the header must be padded when
signing.

A fixed header size of 0x200 is temporary used, however this header size
should be updated so that it is defined by partition manager.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>