partition manager: disable PM when LMA adjustment is enabled
Post-processing of LMA adjustment is incompatible with partition
manager. This has not been discovered in the past but this limitation
is now exposed as upstream samples starts to use sysbuild and adjust
extra images through the use of BUILD_OUTPUT_ADJUST_LMA to locate
extra images in other flash slots.

Therefore disable partition manager when BUILD_OUTPUT_ADJUST_LMA is
enabled.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>